requirements.txt
flake8~=7.0.0
flask~=3.0.3
flask_sqlalchemy~=3.1.1
gunicorn~=22.0.0
PyMySQL~=1.1.1
PyYAML~=6.0.1
requests~=2.32.0
SQLAlchemy~=2.0.29
setuptools>=73.0.1
